dashboard/requirements.txt
streamlit==1.31.1
mysql-connector-python==8.3.0
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1
